requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Machine Learning (lighter versions for Vercel)
tensorflow-cpu==2.15.0
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.4

# HTTP requests
requests==2.31.0

# CORS and middleware
python-multipart==0.0.6

# Development and utilities
python-dotenv==1.0.0
